+++ controllers/accountant/referenceData/street.js
@@ -1,6 +1,10 @@
 const ErrorResponse = require('../../../utils/errorResponse');
 const asyncHandler = require('../../../middleware/async');
 const Street = require('../../../models/accountant/referenceData/Street');
+const Client = require('../../../models/accountant/referenceData/Client');
+const OurFirm = require('../../../models/accountant/referenceData/OurFirm');
+const Supplier = require('../../../models/accountant/referenceData/Supplier');
+const Worker = require('../../../models/accountant/referenceData/Worker');
 
 //@desc   Add a Street
 //@route  POST /api/v1/accountant/street
@@ -88,15 +92,38 @@ exports.getOneStreet = asyncHandler(async (req, res, next) => {
 //@route  DELETE /api/v1/accountant/street/:id
 //@access Private
 exports.deleteStreet = asyncHandler(async (req, res, next) => {
-  const oneStreet = await Street.findByIdAndDelete(req.params.id);
+  const relatedClient = await Client.findOne({ street: req.params.id }, '_id');
+  const relatedOurFirm = await OurFirm.findOne(
+    { street: req.params.id },
+    '_id'
+  );
+  const relatedSupplier = await Supplier.findOne(
+    { street: req.params.id },
+    '_id'
+  );
+  const relatedWorker = await Worker.findOne({ street: req.params.id }, '_id');
 
-  //Check if  exists response
-  if (!oneStreet) {
-    return next(new ErrorResponse('Нет  объекта с данным id', 400));
+  const forbiddenToDelete =
+    relatedClient || relatedOurFirm || relatedSupplier || relatedWorker;
+
+  if (forbiddenToDelete) {
+    return next(
+      new ErrorResponse(
+        'не возможно удалить этот елемент, есть связанные элементы',
+        403
+      )
+    );
+  } else {
+    const oneStreet = await Street.findByIdAndDelete(req.params.id);
+
+    //Check if  exists response
+    if (!oneStreet) {
+      return next(new ErrorResponse('Нет  объекта с данным id', 400));
+    }
+
+    res.status(200).json({
+      success: true,
+      data: {}
+    });
   }
-
-  res.status(200).json({
-    success: true,
-    data: {}
-  });
 });
