+++ controllers/accountant/referenceData/typeOf_Settlement.js
@@ -1,6 +1,10 @@
 const ErrorResponse = require('../../../utils/errorResponse');
 const asyncHandler = require('../../../middleware/async');
 const TypeOf_Settlement = require('../../../models/accountant/referenceData/TypeOf_Settlement');
+const Client = require('../../../models/accountant/referenceData/Client');
+const OurFirm = require('../../../models/accountant/referenceData/OurFirm');
+const Supplier = require('../../../models/accountant/referenceData/Supplier');
+const Worker = require('../../../models/accountant/referenceData/Worker');
 
 //@desc   Add a TypeOf_Settlement
 //@route  POST /api/v1/accountant/type-of-settlement
@@ -90,17 +94,46 @@ exports.getOneTypeOf_Settlement = asyncHandler(async (req, res, next) => {
 //@route  DELETE /api/v1/accountant/type-of-settlement/:id
 //@access Private
 exports.deleteTypeOf_Settlement = asyncHandler(async (req, res, next) => {
-  const oneTypeOf_Settlement = await TypeOf_Settlement.findByIdAndDelete(
-    req.params.id
+  const relatedClient = await Client.findOne(
+    { typeOf_settlement: req.params.id },
+    '_id'
+  );
+  const relatedOurFirm = await OurFirm.findOne(
+    { typeOf_settlement: req.params.id },
+    '_id'
+  );
+  const relatedSupplier = await Supplier.findOne(
+    { typeOf_settlement: req.params.id },
+    '_id'
+  );
+  const relatedWorker = await Worker.findOne(
+    { typeOf_settlement: req.params.id },
+    '_id'
   );
 
-  //Check if  exists response
-  if (!oneTypeOf_Settlement) {
-    return next(new ErrorResponse('Нет  объекта с данным id', 400));
-  }
+  const forbiddenToDelete =
+    relatedClient || relatedOurFirm || relatedSupplier || relatedWorker;
 
-  res.status(200).json({
-    success: true,
-    data: {}
-  });
+  if (forbiddenToDelete) {
+    return next(
+      new ErrorResponse(
+        'не возможно удалить этот елемент, есть связанные элементы',
+        403
+      )
+    );
+  } else {
+    const oneTypeOf_Settlement = await TypeOf_Settlement.findByIdAndDelete(
+      req.params.id
+    );
+
+    //Check if  exists response
+    if (!oneTypeOf_Settlement) {
+      return next(new ErrorResponse('Нет  объекта с данным id', 400));
+    }
+
+    res.status(200).json({
+      success: true,
+      data: {}
+    });
+  }
 });
