+++ controllers/accountant/referenceData/typeOf_TaxPayerOn.js
@@ -1,6 +1,9 @@
 const ErrorResponse = require('../../../utils/errorResponse');
 const asyncHandler = require('../../../middleware/async');
 const TypeOf_TaxPayerOn = require('../../../models/accountant/referenceData/TypeOf_TaxPayerOn');
+const Client = require('../../../models/accountant/referenceData/Client');
+const OurFirm = require('../../../models/accountant/referenceData/OurFirm');
+const Supplier = require('../../../models/accountant/referenceData/Supplier');
 
 //@desc   Add a TypeOf_TaxPayerOn
 //@route  POST /api/v1/accountant/type-of-tax-payer-on
@@ -88,17 +91,41 @@ exports.getOneTypeOf_TaxPayerOn = asyncHandler(async (req, res, next) => {
 //@route  DELETE /api/v1/accountant/type-of-tax-payer-on/:id
 //@access Private
 exports.deleteTypeOf_TaxPayerOn = asyncHandler(async (req, res, next) => {
-  const oneTypeOf_TaxPayerOn = await TypeOf_TaxPayerOn.findByIdAndDelete(
-    req.params.id
+  const relatedClient = await Client.findOne(
+    { taxPayerOn: req.params.id },
+    '_id'
+  );
+  const relatedOurFirm = await OurFirm.findOne(
+    { taxPayerOn: req.params.id },
+    '_id'
+  );
+  const relatedSupplier = await Supplier.findOne(
+    { taxPayerOn: req.params.id },
+    '_id'
   );
 
-  //Check if  exists response
-  if (!oneTypeOf_TaxPayerOn) {
-    return next(new ErrorResponse('Нет  объекта с данным id', 400));
-  }
+  const forbiddenToDelete = relatedClient || relatedOurFirm || relatedSupplier;
 
-  res.status(200).json({
-    success: true,
-    data: {}
-  });
+  if (forbiddenToDelete) {
+    return next(
+      new ErrorResponse(
+        'не возможно удалить этот елемент, есть связанные элементы',
+        403
+      )
+    );
+  } else {
+    const oneTypeOf_TaxPayerOn = await TypeOf_TaxPayerOn.findByIdAndDelete(
+      req.params.id
+    );
+
+    //Check if  exists response
+    if (!oneTypeOf_TaxPayerOn) {
+      return next(new ErrorResponse('Нет  объекта с данным id', 400));
+    }
+
+    res.status(200).json({
+      success: true,
+      data: {}
+    });
+  }
 });
